requirements.txt
# Dependências principais
python-dotenv==1.0.0
loguru==0.7.0
pytz==2023.3
firebase-admin==6.2.0
SQLAlchemy==2.0.29

# Dependências de desenvolvimento
black==24.2.0
flake8==7.0.0
isort==5.13.2
mypy==1.5.1

# Firebase e Armazenamento
google-cloud-storage==2.14.0
google-auth==2.28.1

# Web e API
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
websockets==12.0

# Utilitários
python-dateutil==2.8.2
pydantic==2.6.1
typing-extensions==4.9.0

# Testes
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# Integração com WhatsApp
whatsapp-web.js==1.23.0
qrcode-terminal==0.12.0

# Processamento de Linguagem Natural
nltk==3.8.1
spacy==3.7.2
transformers==4.37.2

# Análise de Dados
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Segurança
cryptography==42.0.2
python-jose==3.3.0
passlib==1.7.4 